user permissions for post deleting

+++ comments/views.py
@@ -1,7 +1,7 @@
 
 from django.contrib import messages
 from django.contrib.contenttypes.models import ContentType
-from django.http import HttpResponseRedirect
+from django.http import HttpResponseRedirect, Http404, HttpResponse
 from django.shortcuts import render, get_object_or_404
 
 # Create your views here.
@@ -11,6 +11,16 @@ from .models import Comment
 
 def comment_delete(request, id):
     obj = get_object_or_404(Comment, id=id)
+    try:
+        obj = Comment.objects.get(id=id)
+    except:
+        raise Http404
+
+    if obj.user != request.user:
+        response = HttpResponse("You do not have permission to delete this.")
+        response.status_code = 403
+        return response
+
     if request.method == "POST":
         parent_obj_url = obj.content_object.get_absolute_url()
         obj.delete()
@@ -23,6 +33,13 @@ def comment_delete(request, id):
 
 def comment_thread(request, id):
     obj = get_object_or_404(Comment, id=id)
+    try:
+        obj = Comment.objects.get(id=id)
+    except:
+        raise Http404
+    
+    if not obj.is_parent:
+        obj = obj.parent
     # print(dir(form))
     # print(form.errors)
     initial_data = {
